add validation exception for mismatched fact table headings
